Fix Processing Array not indicating it is doing work
I had inadvertently left out the setActive call during prior refactoring.

Removed unnecessary wrapper class: turns out the reflection stuff there
before was just because the access specifier for the property we needed
was 'private' until Nov 2019. It is now 'protected' so there was no
reason to have either the reflection or the wrapper class.