Fix potential issue in findAllItemsInInputs, fix indentation
ItemStack doesn't have a hashing function, so HashSet is likely to incorrectly combine items that are not the same by our definition.
Instead, make use of a custom hashing strategy we implemented in GTCE for the Multi Smelter.

Indendation corrected from mixed tabs/spaces to "Smart Tabs" which indents with tabs and aligns within an indent level using spaces.