Fix PA computing chanced outputs incorrectly
A longstanding bug with chanced outputs, arising from improperly using
the machine tier as the number of times to apply the chance function's
bonus per tier, was fixed in GTNE 1.18.1.

The PA overrides the relevant fixed method from AbstractRecipeLogic,
causing a regression to the incorrect older behavior.

The PA's override now uses the corrected behavior.